Changed 'int id' => 'long id'.

+++ bond_risk/objects/calendar.c
@@ -1,7 +1,7 @@
 // Copyright (c) Steve Evans 2010
 // schevans@users.sourceforge.net
 // This code is released under the BSD licence. For details see $APP_ROOT/LICENCE
-//
+// 
 
 // Please see the comment at the top of bond_risk/gen/objects/calendar_macros.h
 //  to see what's in-scope.
@@ -12,9 +12,9 @@
 
 using namespace std;
 
-int calculate_calendar( obj_loc_t obj_loc, int id )
+int calculate_calendar( obj_loc_t obj_loc, long id )
 {
-    // Does nothing.
+    // Write me.
 
     return 1;
 }
